==> notebooks/02_data_cleaning.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

df= pd.read_csv('../data/raw/output.csv')

### Traitement des variables inutiles

# 📌 Suppression des variables inutiles
df_clean = df.drop(columns=["date", "street", "statezip", "country"])

# 📌 Encodage de la variable 'city' par le prix moyen
city_mean_price = df_clean.groupby("city")["price"].mean()

# Création d'une nouvelle variable 'city_mean_price'
df_clean["city_mean_price"] = df_clean["city"].map(city_mean_price)

# Suppression de la colonne 'city' (car remplacée)
df_clean = df_clean.drop(columns=["city"])

df_clean.head()



# On verifie s'il y a bien 44 valeurs uniques pour 'city_mean_price'
# et on affiche leur distribution

variable = 'city_mean_price' 
unique_values = df_clean[variable].unique()

print(len(unique_values))
print(df_clean[variable].value_counts().sort_values())


### Verifications de la plage de la variable binaire waterfront

# Afficher les valeurs et leurs fréquences
print("Distribution de 'waterfront':")
print(df_clean['waterfront'].value_counts().sort_index())

### Les variables conditions et view sont des notes, on verifie leurs distributions pour s'assurer qu'il n y apas de valeurs aberrantes

# Afficher les valeurs et leurs fréquences
print("Distribution de 'condition':")
print(df_clean['condition'].value_counts().sort_index())

# Afficher les valeurs et leurs fréquences
print("Distribution de 'view':")
print(df_clean['view'].value_counts().sort_index())

### Verifions aussi avec les variables bedrooms, bathrooms, et floors

print("Distribution de 'bedrooms':")
print(df_clean['bedrooms'].value_counts().sort_index())

print("Distribution de 'bathrooms':")
print(df_clean['bathrooms'].value_counts().sort_index())


print("Distribution de 'floors':")
print(df_clean['floors'].value_counts().sort_index())

### Il y a des maisons à 0 et plus de 5 etages, on verifie si le prix et la surface habitable de ces maisons sont coherants

# Filtrer les maisons avec 0 ou 8 salles de bain
extreme_bath = df_clean[df_clean['bathrooms'].isin([0.00, 8.00])]
print(extreme_bath[['bathrooms','price', 'sqft_living', 'bedrooms']])

### Verifions les variables des maisons sans salles de bain

no_bath = df_clean[df_clean['bathrooms'] == 0.00]
no_bath.head()

### Beaucoup d'incoherances pour ces 2 maisons donc, on fait le choix de les supprimer.

## Suppression des maisons avec 0 salles de bain et reoncstruction du dataframe

df = df[df['bathrooms'] > 0]

df_clean = df.drop(columns=["date", "street", "statezip", "country"])
city_mean_price = df_clean.groupby("city")["price"].mean()
df_clean["city_mean_price"] = df_clean["city"].map(city_mean_price)
df_clean = df_clean.drop(columns=["city"])

df_clean.head()

no_bath = df_clean[df_clean['bathrooms'] == 0.00]
no_bath.head()

### Traitement des variables yr_built et yr_renovated

print("Distribution de 'yr_built':")
print(df_clean['yr_built'].value_counts().sort_index())

print("Distribution de 'yr_renovated':")
print(df_clean['yr_renovated'].value_counts().sort_index())

#Calculer l'âge de la maison par rapport à 2018
df_clean["house_age"] = 2018 - df["yr_built"]

#Variable binaire : rénovée ou non
df_clean["renovated"] = np.where(df["yr_renovated"] > 0, 1, 0)

#Âge depuis la dernière rénovation (si pas rénovée → âge total)
df_clean["age_since_renov"] = np.where(
    df["yr_renovated"] > 0,
    2018 - df["yr_renovated"],
    df_clean["house_age"]
)

#Supprimer les colonnes d'origine
df_clean = df_clean.drop(columns=["yr_built", "yr_renovated"])

df_clean.head()


### Identification des variables ayant une correlation superieure à 0.8


#Calcul de la matrice de corrélation
corr_matrix = df_clean.corr(numeric_only=True)

#On récupère uniquement la partie supérieure (pour éviter les doublons)
upper = corr_matrix.where(np.triu(np.ones(corr_matrix.shape), k=1).astype(bool))

#Sélection des paires de variables corrélées ≥ 0.8
high_corr_pairs = [
    (col, row, upper.loc[row, col])
    for col in upper.columns
    for row in upper.index
    if abs(upper.loc[row, col]) >= 0.8
]

print("Variables très corrélées entre elles (|corr| ≥ 0.8) :\n")
for var1, var2, corr_val in high_corr_pairs:
    print(f"{var1}  <-->  {var2}   | corr = {corr_val:.2f}")


df_clean = df_clean.drop(columns=["sqft_above"])

df_clean.head()

### Sauvegarde du dataset pour les prochaines etapes

import os
from pathlib import Path


# Chemin relatif pour la sauvegarde
output_path = Path("../data/processed/df_clean.pkl")  # Je recommande 'processed' plutôt qu'external

# Sauvegarde en pickle (conservation des types)
df_clean.to_pickle(output_path)
