==> notebooks/03_feature_ingineering.ipynb <==
import pandas as pd
from pathlib import Path
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
import joblib
import os

# Chargement des données nettoyées
df_clean = pd.read_pickle("../data/processed/df_clean.pkl")

# Vérification
print(f"Shape du dataset chargé : {df_clean.shape}")
display(df_clean.head())

# Définir la cible et les features
TARGET = "price"
X = df_clean.drop(columns=[TARGET])
y = df_clean[TARGET]

#Séparer Train / Test
X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, random_state=42
)

#Standardiser les features numériques
df_cols = X_train.columns

scaler = StandardScaler()
X_train_scaled = X_train.copy()
X_test_scaled = X_test.copy()

X_train_scaled[df_cols] = scaler.fit_transform(X_train[df_cols])
X_test_scaled[df_cols] = scaler.transform(X_test[df_cols])

#Sauvegarder les fichiers
os.makedirs("../data/processed", exist_ok=True)

X_train_scaled.to_csv("../data/processed/X_train.csv", index=False)
X_test_scaled.to_csv("../data/processed/X_test.csv", index=False)
y_train.to_csv("../data/processed/y_train.csv", index=False)
y_test.to_csv("../data/processed/y_test.csv", index=False)

# Sauvegarder le scaler entraîné
joblib.dump(scaler, "../data/processed/scaler.pkl")

print("✅ Données et scaler sauvegardés dans ../data/processed/")
